Refactor project as package
