+++ sniper_main/notifier.py
@@ -2,7 +2,7 @@ from telegram import Bot
 import smtplib
 import ssl
 from email.mime.text import MIMEText
-from config import Config
+from .config import Config
 
 # Load configuration once
 cfg = Config.from_json()
