+++ sniper_main/steal_engine.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 
 from decimal import Decimal
 
-from config import Config
-from db import get_last_30d_avg
-from models import FlightOffer
+from .config import Config
+from .db import get_last_30d_avg
+from .models import FlightOffer
 
 
 def is_steal(offer: FlightOffer, cfg: Config) -> bool:
