+++ sniper_main/tests/test_aggregator.py
@@ -1,19 +1,22 @@
 import os
-import sys
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
-
 from datetime import datetime, timedelta
 import sqlite3
 
 import pytest
 
-from db import init_db
-import aggregator
+from sniper_main.db import init_db
+from sniper_main import aggregator
 
 
 def test_aggregate_30_days_no_gaps(tmp_path):
     db_file = tmp_path / "test.db"
-    schema_path = os.path.join(os.path.dirname(__file__), "..", "..", "schema.sql")
+    schema_path = os.path.join(
+        os.path.dirname(__file__),
+        "..",
+        "..",
+        "sniper-main",
+        "schema.sql",
+    )
     init_db(str(db_file), schema_path=schema_path)
 
     conn = sqlite3.connect(db_file)
