+++ sniper_main/tests/test_config.py
@@ -1,9 +1,7 @@
 import os
-import sys
 import json
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
-from config import Config
+from sniper_main.config import Config
 
 
 def test_load_config(tmp_path):
