+++ sniper_main/tests/test_steal_engine.py
@@ -1,11 +1,8 @@
-import os
-import sys
 from dataclasses import dataclass
 from decimal import Decimal
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
-from config import Config
-import steal_engine
+from sniper_main.config import Config
+from sniper_main import steal_engine
 
 
 @dataclass(slots=True)
